Fix compilation in case if key is void*
If key type is `void*`, wrong `operator()` is chosen because of the dummy argument that is used for SFINAE. As a result the key comparison adapter invokes the user's predicate with only one argument, which fails to compile.

Use a more distict type for SFINAE in the dummy arguments. Fixes ticket #12745.